fix: api生成以及整合hook
