+++ src/input/props.ts
@@ -4,7 +4,6 @@
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
 
-import { NOOP } from '@/shared';
 import { TdInputProps } from './type';
 import { PropType } from 'vue';
 
@@ -40,7 +39,7 @@ export default {
   },
   /** 是否可清空 */
   clearable: Boolean,
-  /** 光标颜色，默认颜色值 #0052d9 */
+  /** 光标颜色 */
   cursorColor: {
     type: String,
     default: '#0052d9',
@@ -155,9 +154,6 @@ export default {
   onClear: Function as PropType<TdInputProps['onClear']>,
   /** 获得焦点时触发 */
   onFocus: Function as PropType<TdInputProps['onFocus']>,
-  /** 字数超出限制时触发 */
-  onValidate: {
-    type: Function as PropType<TdInputProps['onValidate']>,
-    default: NOOP,
-  },
+  /** 【暂不支持】字数超出限制时触发 */
+  onValidate: Function as PropType<TdInputProps['onValidate']>,
 };
